revert(pi): drop the shared.ts extraction — runners stay self-contained
The switchboard is the single decision point; a shared layer between it and
the runners is a place decisions could creep into. index.ts, task.ts, and
subagent.ts return to the base branch verbatim; the typed effort in
gateway.ts/types.ts stays.

Generated-By: PostHog Code
Task-Id: fafc230d-6f14-4e4d-9462-0e7f18a1eec1
