+++ src/lib/agent/runner/harness/pi/index.ts
@@ -16,22 +16,17 @@ import fs from 'fs';
 import path from 'path';
 import { getUI } from '@ui';
 import { getLogFilePath, logToFile } from '@utils/debug';
-import { Harness, WIZARD_REMARK_EVENT_NAME } from '@lib/constants';
+import {
+  Harness,
+  WIZARD_REMARK_EVENT_NAME,
+  WIZARD_USER_AGENT,
+} from '@lib/constants';
 import { analytics } from '@utils/analytics';
 import { AgentErrorType } from '@lib/agent/agent-interface';
-import { REMARK_INSTRUCTION } from '@lib/agent/signals';
+import { AgentSignals, REMARK_INSTRUCTION } from '@lib/agent/signals';
 import { AgentOutputSignals } from '@lib/agent/output-signals';
 import { getWizardCommandments } from '@lib/agent/commandments';
-import {
-  captureAgentUsage,
-  classifyRunError,
-  connectPostHogMcp,
-  createHermeticSession,
-  piCodingToolFactories,
-  resolveGatewayModel,
-  startRunClock,
-  watchSession,
-} from './shared';
+import { buildGatewayProvider, GATEWAY_PROVIDER } from './gateway';
 import type {
   AgentResult,
   AgentHarness,
@@ -99,6 +94,113 @@ function piMcpContext(boot: BootstrapResult, instructions?: string): string {
   ].join('\n');
 }
 
+/**
+ * The ONLY environment variables pi's tool subprocesses (bash → npm/pip/…) are
+ * allowed to see. Everything else — every secret (POSTHOG_PERSONAL_API_KEY,
+ * ANTHROPIC_*, AWS_*), every ambient credential, the parent process's whole env
+ * — is dropped before a child is spawned. pi's own gateway auth is programmatic
+ * (the access token never lives in env), so a minimal env costs the agent
+ * nothing while closing the leak that exposed the key before. Kept to what a
+ * package manager genuinely needs to run.
+ */
+const ALLOWED_SUBPROCESS_ENV_KEYS = [
+  'PATH',
+  'HOME',
+  'SHELL',
+  'USER',
+  'LOGNAME',
+  'TMPDIR',
+  'TMP',
+  'TEMP',
+  'TERM',
+  'LANG',
+  'LC_ALL',
+  'LC_CTYPE',
+  'NODE_EXTRA_CA_CERTS',
+  'SSL_CERT_FILE',
+  'SSL_CERT_DIR',
+  'HTTP_PROXY',
+  'HTTPS_PROXY',
+  'NO_PROXY',
+  'http_proxy',
+  'https_proxy',
+  'no_proxy',
+];
+
+/** A fresh subprocess env holding only the allowlisted keys present in process.env. */
+export function buildScrubbedEnv(): NodeJS.ProcessEnv {
+  const env: NodeJS.ProcessEnv = {};
+  for (const key of ALLOWED_SUBPROCESS_ENV_KEYS) {
+    const value = process.env[key];
+    if (value !== undefined) env[key] = value;
+  }
+  return env;
+}
+
+/**
+ * Tag a tool with an execution mode (mutates + returns it). Read-only tools are
+ * `parallel` so a single turn that batches independent reads/searches runs them
+ * at once; mutating/install tools are `sequential` so a batch never races writes
+ * or concurrent installs. pi-agent-core runs a batch in parallel only when no
+ * tool in it is `sequential`.
+ */
+export function withMode<T>(tool: T, mode: 'sequential' | 'parallel'): T {
+  (tool as { executionMode?: 'sequential' | 'parallel' }).executionMode = mode;
+  return tool;
+}
+
+/** Pull plain text out of a pi AgentMessage (content is text/image blocks). */
+export function extractText(message: unknown): string {
+  const content = (message as { content?: unknown })?.content;
+  if (typeof content === 'string') return content;
+  if (Array.isArray(content)) {
+    return content
+      .filter((c): c is { type: string; text: string } => {
+        const block = c as { type?: string; text?: unknown };
+        return block?.type === 'text' && typeof block.text === 'string';
+      })
+      .map((c) => c.text)
+      .join('');
+  }
+  return '';
+}
+
+/**
+ * Surface `[DASHBOARD_URL]` / `[NOTEBOOK_URL]` markers the agent prints (after
+ * the MCP creates them) into the outro link, mirroring the anthropic path's
+ * signal parsing (#9). The marker carries the URL the MCP returned.
+ */
+export function applyOutroMarkers(textBlock: string): void {
+  const markers: Array<[string, (url: string) => void]> = [
+    [AgentSignals.DASHBOARD_URL, (url) => getUI().setDashboardUrl(url)],
+    [AgentSignals.NOTEBOOK_URL, (url) => getUI().setNotebookUrl(url)],
+  ];
+  for (const [marker, apply] of markers) {
+    const idx = textBlock.indexOf(marker);
+    if (idx === -1) continue;
+    const url = textBlock
+      .slice(idx + marker.length)
+      .trim()
+      .split(/\s/)[0];
+    if (url) apply(url);
+  }
+}
+
+/**
+ * The text of the last `[STATUS] …` line in a block, if any. Last wins so the
+ * spinner shows the most recent action when a turn prints several.
+ */
+export function lastStatusLine(textBlock: string): string | undefined {
+  let status: string | undefined;
+  for (const line of textBlock.split('\n')) {
+    const idx = line.indexOf(AgentSignals.STATUS);
+    if (idx !== -1) {
+      status = line.slice(idx + AgentSignals.STATUS.length).trim();
+    }
+  }
+  return status || undefined;
+}
+
 /** Cap on completion-guard re-prompts while tasks remain open (see the run loop). */
 const MAX_CONTINUE_NUDGES = 20;
 
@@ -127,8 +229,19 @@ export const piBackend: AgentHarness = {
     spinner.start(config.spinnerMessage ?? 'Customizing your PostHog setup...');
 
     // Same `agent completed`/`agent aborted` shape as anthropic.
+    const startTime = Date.now();
     const signals = new AgentOutputSignals();
-    const runDurations = startRunClock();
+    let assistantTurns = 0;
+    // Tool calls across the whole run. Zero means the agent only ever produced
+    // text and never acted — a no-op that leaves the project untouched.
+    let toolCalls = 0;
+    const runDurations = () => {
+      const durationMs = Date.now() - startTime;
+      return {
+        duration_ms: durationMs,
+        duration_seconds: Math.round(durationMs / 1000),
+      };
+    };
     const captureAborted = () =>
       analytics.wizardCapture('agent aborted', {
         ...runDurations(),
@@ -136,17 +249,41 @@ export const piBackend: AgentHarness = {
       });
 
     try {
-      const sdk = await import('@earendil-works/pi-coding-agent');
-      const { getAgentDir } = sdk;
+      const {
+        createAgentSession,
+        DefaultResourceLoader,
+        SessionManager,
+        AuthStorage,
+        ModelRegistry,
+        getAgentDir,
+        createLsToolDefinition,
+        createFindToolDefinition,
+        createGrepToolDefinition,
+        createBashToolDefinition,
+        createReadToolDefinition,
+        createEditToolDefinition,
+        createWriteToolDefinition,
+      } = await import('@earendil-works/pi-coding-agent');
 
-      const gateway = resolveGatewayModel(sdk, boot, modelId);
-      if (!gateway) {
+      // the claude-agent-sdk path. The provider spec is shared with the
+      // orchestrator's per-task sessions (gateway.ts).
+      const { provider, caps, gatewayUrl } = buildGatewayProvider({
+        gatewayUrl: boot.credentials.host.gatewayUrl,
+        accessToken: boot.credentials.accessToken,
+        wizardMetadata: boot.wizardMetadata,
+        wizardFlags: boot.wizardFlags,
+        modelId,
+      });
+      const registry = ModelRegistry.inMemory(AuthStorage.create());
+      registry.registerProvider(GATEWAY_PROVIDER, provider as never);
+
+      const model = registry.find(GATEWAY_PROVIDER, modelId);
+      if (!model) {
         return {
           error: AgentErrorType.API_ERROR,
           message: 'pi: gateway model could not be resolved',
         };
       }
-      const { registry, model, caps, gatewayUrl } = gateway;
 
       // System prompt = wizard commandments. Skip project context files /
       // user extensions / skills so the run is hermetic; skills discovery is a
@@ -177,13 +314,48 @@ export const piBackend: AgentHarness = {
       const { prewarmYaraScanner } = await import('@lib/yara-hooks');
       void prewarmYaraScanner();
 
-      // Real PostHog MCP (#10), best-effort; the security factory is always first.
+      // Wire the real PostHog MCP into pi (#10): load pi's MCP adapter and point
+      // it at the hosted MCP the anthropic path uses, so dashboards/insights are
+      // created through the sanctioned MCP. Best-effort — if it can't load or
+      // connect, the run continues (minus the dashboard step) rather than failing
+      // the whole integration. The security factory is always first.
       const extensionFactories = [security.factory] as Array<
         (pi: unknown) => void
       >;
-      const mcp = await connectPostHogMcp(sdk, boot, '[pi]');
-      if (mcp.extensionFactory) extensionFactories.push(mcp.extensionFactory);
-      const mcpCleanup = mcp.cleanup;
+      let mcpCleanup: (() => void) | undefined;
+      let mcpInstructions: string | undefined;
+      try {
+        const { setupPostHogMcp } = await import('./mcp');
+        const mcp = await setupPostHogMcp({
+          agentDir: getAgentDir(),
+          mcpUrl: boot.credentials.host.mcpUrl,
+          accessToken: boot.credentials.accessToken,
+          userAgent: WIZARD_USER_AGENT,
+        });
+        extensionFactories.push(mcp.extensionFactory);
+        mcpCleanup = mcp.cleanup;
+        mcpInstructions = mcp.instructions;
+      } catch (err) {
+        logToFile(`[pi] PostHog MCP setup skipped: ${String(err)}`);
+      }
+
+      const resourceLoader = new DefaultResourceLoader({
+        cwd: session.installDir,
+        agentDir: getAgentDir(),
+        systemPrompt:
+          getWizardCommandments() +
+          '\n' +
+          PI_RUNTIME_NOTES +
+          '\n' +
+          piMcpContext(boot, mcpInstructions),
+        noExtensions: true,
+        noSkills: true,
+        noContextFiles: true,
+        noPromptTemplates: true,
+        noThemes: true,
+        extensionFactories,
+      });
+      await resourceLoader.reload();
 
       // Wizard capabilities as custom tools (pi has no MCP): skill
       // discovery/install + fenced .env edits, same names as the MCP server so
@@ -196,21 +368,31 @@ export const piBackend: AgentHarness = {
       const { createDispatchAgentTool } = await import('./subagent');
       // Created once so the run loop can read the store for the completion guard.
       const wizardTaskTools = createWizardPiTaskTools();
-      // The one env-scrubbed bash is shared with the subagent so the lockdown is inherited.
-      const coding = piCodingToolFactories(sdk, session.installDir);
-      const scrubbedBash = coding.bash();
+      // The one bash the agent (and its subagents) may use: every subprocess it
+      // spawns gets a scrubbed env, so no secret or ambient variable reaches an
+      // `npm install`. Shared with the subagent so the lockdown is inherited.
+      const scrubbedBash = withMode(
+        createBashToolDefinition(session.installDir, {
+          spawnHook: (ctx) => ({ ...ctx, env: buildScrubbedEnv() }),
+        }),
+        'sequential',
+      );
 
       const customTools = [
-        coding.read(),
-        coding.edit(),
-        coding.write(),
+        // Built-ins re-registered explicitly. `noTools: 'builtin'` disables pi's
+        // defaults so we can supply the env-scrubbed bash above; read/edit/write
+        // are the stock definitions. Reads run in parallel so a batched turn of
+        // independent reads executes at once; edit/write/bash stay sequential.
+        withMode(createReadToolDefinition(session.installDir), 'parallel'),
+        withMode(createEditToolDefinition(session.installDir), 'sequential'),
+        withMode(createWriteToolDefinition(session.installDir), 'sequential'),
         scrubbedBash,
         // Native ls/find/grep so the agent explores with proper tools instead
         // of fence-blocked `bash {ls/find}` (the profiled retry-spirals came
         // from this gap). Parallel — exploration batches cleanly.
-        coding.ls(),
-        coding.find(),
-        coding.grep(),
+        withMode(createLsToolDefinition(session.installDir), 'parallel'),
+        withMode(createFindToolDefinition(session.installDir), 'parallel'),
+        withMode(createGrepToolDefinition(session.installDir), 'parallel'),
         ...createWizardPiTools({
           workingDirectory: session.installDir,
           skillsBaseUrl: boot.skillsBaseUrl,
@@ -229,32 +411,85 @@ export const piBackend: AgentHarness = {
           agentDir: getAgentDir(),
           securityFactory: security.factory as (pi: unknown) => void,
           bashTool: scrubbedBash,
-          sdk,
+          sdk: { createAgentSession, DefaultResourceLoader, SessionManager },
         }),
       ];
 
-      const agentSession = await createHermeticSession(sdk, {
-        cwd: session.installDir,
-        systemPrompt:
-          getWizardCommandments() +
-          '\n' +
-          PI_RUNTIME_NOTES +
-          '\n' +
-          piMcpContext(boot, mcp.instructions),
-        extensionFactories,
+      const { session: agentSession } = await createAgentSession({
         model,
-        registry,
+        modelRegistry: registry,
         // Reasoning effort from the switchboard capability matrix (undefined =
         // pi's default). Sent as `reasoning_effort` for openai-completions.
         thinkingLevel: caps.thinkingLevel,
+        cwd: session.installDir,
+        sessionManager: SessionManager.inMemory(session.installDir),
+        resourceLoader,
+        // Disable the default built-in tools; `customTools` re-registers
+        // read/edit/write + an env-scrubbed bash, so no subprocess inherits the
+        // host env. Custom + extension tools stay enabled.
+        noTools: 'builtin',
         customTools,
       });
 
-      // counts drive the no-progress guard below.
-      const { counts, unsubscribe } = watchSession(agentSession, {
-        tag: '[pi]',
-        spinner,
-        signals,
+      // Fire the extension lifecycle — what interactive mode does via
+      // rebindCurrentSession. createAgentSession builds the session but does not
+      // emit session_start on its own, and the MCP adapter connects on that
+      // event; without this its tools report "MCP not initialized".
+      await agentSession.bindExtensions({});
+
+      // Map pi events onto the run spinner + the log file, mirroring the
+      // anthropic path's log shape (assistant turns + tool I/O) and driving the
+      // single run spinner with one stable status at a time (no overlap).
+      const unsubscribe = agentSession.subscribe((event) => {
+        switch (event.type) {
+          case 'message_end': {
+            // User prompts also emit message_end; only assistant turns count.
+            if ((event.message as { role?: string })?.role !== 'assistant') {
+              break;
+            }
+            assistantTurns += 1;
+            const assistant = extractText(event.message).trim();
+            if (assistant) {
+              logToFile(`[pi] assistant: ${assistant.slice(0, 1000)}`);
+              applyOutroMarkers(assistant);
+              // Surface [STATUS] lines into the live spinner + status history,
+              // mirroring the anthropic path — pi otherwise drops them.
+              const statusText = lastStatusLine(assistant);
+              if (statusText) {
+                getUI().pushStatus(statusText);
+                spinner.message(statusText);
+              }
+              for (const line of assistant.split('\n')) signals.push(line);
+            }
+            break;
+          }
+          case 'tool_execution_start': {
+            toolCalls += 1;
+            const args = JSON.stringify(event.args ?? {}).slice(0, 200);
+            logToFile(`[pi] → ${event.toolName} ${args}`);
+            // Don't surface raw tool names in the spinner — the anthropic path
+            // doesn't, and it reads as noise. The Task panel (syncTodos) is the
+            // visible progress, matching the anthropic presentation.
+            break;
+          }
+          case 'tool_execution_end': {
+            if (event.isError) {
+              logToFile(
+                `[pi] ✗ ${event.toolName}: ${String(event.result).slice(
+                  0,
+                  300,
+                )}`,
+              );
+            }
+            break;
+          }
+          case 'agent_end': {
+            logToFile(`[pi] agent_end (willRetry=${String(event.willRetry)})`);
+            break;
+          }
+          default:
+            break;
+        }
       });
 
       try {
@@ -305,17 +540,14 @@ export const piBackend: AgentHarness = {
       // pi ends a run on any tool-call-less turn, so guard against a hollow
       // success reaching the outro (nothing done, or stopped mid-plan).
       const openTasks = hasOpenTasks(wizardTaskTools.store);
-      const failure = completionFailure({
-        toolCalls: counts.toolCalls,
-        openTasks,
-      });
+      const failure = completionFailure({ toolCalls, openTasks });
       if (failure === AgentErrorType.NO_PROGRESS) {
         spinner.stop('Agent made no changes');
         logToFile(
-          `[pi] no progress: ${counts.assistantTurns} assistant turn(s), 0 tool calls`,
+          `[pi] no progress: ${assistantTurns} assistant turn(s), 0 tool calls`,
         );
         analytics.wizardCapture('agent no progress', {
-          assistant_turns: counts.assistantTurns,
+          assistant_turns: assistantTurns,
         });
         captureAborted();
         return { error: failure };
@@ -352,22 +584,32 @@ export const piBackend: AgentHarness = {
         logToFile(`[pi] .posthog-events.json cleanup skipped: ${String(err)}`);
       }
 
-      captureAgentUsage({
-        tag: '[pi]',
-        modelId,
-        counts,
-        durations: runDurations(),
-        stats: agentSession.getSessionStats(),
+      const stats = agentSession.getSessionStats();
+      analytics.wizardCapture('agent completed', {
+        ...runDurations(),
+        model: modelId,
+        num_turns: assistantTurns,
+        // API-reported tokens only; no total_cost_usd — the API returns no
+        // cost, and $ai_generation already prices the run authoritatively.
+        input_tokens: stats.tokens.input,
+        output_tokens: stats.tokens.output,
+        cache_creation_input_tokens: stats.tokens.cacheWrite,
+        cache_read_input_tokens: stats.tokens.cacheRead,
       });
       spinner.stop(config.successMessage ?? 'PostHog integration complete');
       return {};
     } catch (err) {
-      const { error, message } = classifyRunError(err);
+      const message = err instanceof Error ? err.message : String(err);
       logToFile(`[pi] run error: ${message}`);
       spinner.stop(config.errorMessage ?? `${config.integrationLabel} failed`);
       getUI().log.error(`pi backend error: ${message}`);
       captureAborted();
-      return { error, message };
+
+      const lower = message.toLowerCase();
+      if (lower.includes('rate limit') || lower.includes('429')) {
+        return { error: AgentErrorType.RATE_LIMIT, message };
+      }
+      return { error: AgentErrorType.API_ERROR, message };
     }
   },
